refactor: 메인페이지 구조 리팩토링
